requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.1
django-enumchoicefield==3.0.1
djangorestframework==3.14.0
exceptiongroup==1.1.2
executing==1.2.0
factory-boy==3.2.1
Faker==18.7.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
